型を合わせた

+++ src/config/server_context.cpp
@@ -42,7 +42,7 @@ const LocationContext &ServerContext::SearchLocation(
     const std::string &path) const {
   std::map<std::string, LocationContext>::const_iterator it = location_.begin();
   std::map<std::string, LocationContext>::const_iterator ret = location_.end();
-  int ret_len = 0;
+  long unsigned int ret_len = 0;
   for (; it != location_.end(); it++) {
     if (it->first[0] == '=' && it->first.substr(2) == path) {
       return it->second;
